Fixed overflow

+++ FindLogarithm/MyPair.java
@@ -3,7 +3,7 @@ package FindLogarithm;
 public class MyPair {
     private final int u;
     private final int v;
-    private int resultX = 0;
+    private long resultX = 0;
 
     public MyPair(int uValue, int vValue) {
         u = uValue;
@@ -18,11 +18,11 @@ public class MyPair {
         return v;
     }
 
-    public void setResult(int result) {
+    public void setResult(long result) {
         resultX = result;
     }
 
-    public int getResult() {
+    public long getResult() {
         return resultX;
     }
 }
